requirements.txt
gradio
pandas
plotly
python-dotenv
google-generativeai
Pillow
